elt_kaggle_function/requirements.txt
kaggle==1.5.16
google-cloud-storage==2.10.0
pandas==2.1.0
